Add files via upload
Today I finished doing the pages and the CSS, and though I have some errors that I have to ask, I have finished two pages, because I decided to eliminate the second one (About), that I had not had the chance to begin organizing. Also, I looked for a video, downloaded it and put it into the html.